+++ src/group_operations/multiplication_operation_abelian.jl
@@ -52,9 +52,9 @@ _doc_diff_conjugate_abelmul = """
     diff_conjugate!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, Y, g, h, X)
 
 Compute the differential of the conjugate ``c_g(h) = g$(_math(:∘))h$(_math(:∘))g^{-1} = ghg^{-1}``,
-which simplifies for an [`AbelianMultiplicationGroupOperation`](@ref) to ``D(c_g(h))[X] = X``.
+which simplifies for an [`AbelianMultiplicationGroupOperation`](@ref) to ``$(_math(:d))(c_g(h))[X] = X``.
 
-This can be computed in-place of `Y` if `Y` is `mutable`.
+This can be computed in-place of `Y`.
 """
 
 @doc "$(_doc_diff_conjugate_abelmul)"
@@ -88,13 +88,13 @@ _doc_diff_inv_abelmult = """
     diff_inv(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, X)
     diff_inv!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, Y, g, X)
 
-Compute the value of the differential ``Dι_{$(_math(:G))}(g)[X]`` of the inversion ``ι_{$(_math(:G))}(g) := g^{-1}`` at ``X ∈ 𝔤``
+Compute the value of the differential ``$(_math(:d))ι_{$(_math(:G))}(g)[X]`` of the inversion ``ι_{$(_math(:G))}(g) := g^{-1}`` at ``X ∈ 𝔤``
 in the [`LieAlgebra`](@ref) ``𝔤`` of the [`LieGroup`](@ref) `G`.
 
-In the abelian case, the computation simplifies to
+In the Abelian case, the computation simplifies to
 
 ```math
-Dι_{$(_math(:G))}(g)[X] = -gXg^{-1} = -X.
+$(_math(:d))ι_{$(_math(:G))}(g)[X] = -gXg^{-1} = -X.
 ```
 
 This can be computed in-place of `Y` if `Y` is `mutable`.
@@ -109,9 +109,14 @@ _doc_diff_left_compose_abelmult = """
     diff_left_compose(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, h, X)
     diff_left_compose!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, Y, g, h, X)
 
-Compute the differential of the left group multiplication ``λ_g(h) = g$(_math(:∘))h``.
+Compute the differential of the group operation ``g$(_math(:∘))h``, on an [`AbstractLieGroup`](@ref) `G`
+with respect to its first (left) argument `g`.
 
-Due to differences in the representation of some abelian Lie groups, this method wraps a concrete implementation of a specific abelian LieGroup with inputs of type `AbstractArray{<:Any,0}` and supports in-place computation.
+Another interpretation is to consider a function where we do a fixed multiplication from the right with `h`.
+i..e. the right group multiplication function ``ρ_h(g) = g$(_math(:∘))h``.
+
+The differential simplifies for an [`AbelianMultiplicationGroupOperation`](@ref) to the identity, i.e.
+``$(_math(:d))ρ_h(g)[X] = X``.
 
 This can be computed in-place of `Y` if `Y` is `mutable`.
 """ #
@@ -130,19 +135,22 @@ end
 function diff_left_compose!(
         G::LieGroup{𝔽, <:AbelianMultiplicationGroupOperation}, Y, g, h, X
     ) where {𝔽}
-    return copyto!(LieAlgebra(G), Y, diff_left_compose(G, g, h, X))
+    return copyto!(LieAlgebra(G), Y, X)
 end
 
 _doc_diff_right_compose_abelmult = """
     diff_right_compose(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, g, h, X)
     diff_right_compose!(G::LieGroup{𝔽,<:AbelianMultiplicationGroupOperation}, Y, g, h, X)
 
-Compute the differential of the right group multiplication ``ρ_g(h) = h$(_math(:∘))g``.
+Compute the differential of the group operation ``g$(_math(:∘))h``, on an [`AbstractLieGroup`](@ref) `G`
+with respect to its second (right) argument `h`.
 
-Due to differences in the representation of some abelian Lie groups, this method wraps a concrete implementation of a specific abelian LieGroup with inputs of type `AbstractArray{<:Any,0}` and supports in-place computation.
+Another interpretation is to consider a function where we do a fixed multiplication from the left with `g`.
+i..e. the left group multiplication function ``λ_g(h) = g$(_math(:∘))h``.
 
-This can be computed in-place of `Y` if `Y` is `mutable`.
-""" #
+The differential simplifies for an [`AbelianMultiplicationGroupOperation`](@ref) to the identity, i.e.
+``$(_math(:d))λ_g(h)[X] = X``.
+"""
 
 @doc "$(_doc_diff_right_compose_abelmult)"
 function diff_right_compose(
@@ -158,7 +166,7 @@ end
 function diff_right_compose!(
         G::LieGroup{𝔽, <:AbelianMultiplicationGroupOperation}, Y, g, h, X
     ) where {𝔽}
-    return copyto!(LieAlgebra(G), Y, diff_right_compose(G, g, h, X))
+    return copyto!(LieAlgebra(G), Y, X)
 end
 
 _doc_exp_abelmult = """
@@ -229,15 +237,11 @@ function Base.inv(
 end
 
 @doc "$(_doc_inv_abelmult)"
-function inv!(G::LieGroup{𝔽, <:AbelianMultiplicationGroupOperation}, h, g) where {𝔽}
+inv!(::LieGroup{𝔽, <:AbelianMultiplicationGroupOperation}, h, g) where {𝔽}
+
+function _inv!(G::LieGroup{𝔽, <:AbelianMultiplicationGroupOperation}, h, g) where {𝔽}
     copyto!(h, inv(G, g))
     return h
-end
-
-function inv!(
-        G::LieGroup{𝔽, O}, g, ::Identity{O}
-    ) where {𝔽, O <: AbelianMultiplicationGroupOperation}
-    return identity_element!(G, g)
 end
 
 function inv_left_compose(